@@ -1,18 +1,18 @@
 package com.pt.ua.boardgameshub.domain.jpa_domain;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.Table;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
 
 import com.pt.ua.boardgameshub.controller.request_body.GameRequest;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
 
 import java.util.HashSet;
 import java.util.Set;
