@@ -1,16 +1,16 @@
 package com.pt.ua.boardgameshub.domain.jpa_domain;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import javax.persistence.Table;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
 
 import com.pt.ua.boardgameshub.controller.request_body.DeveloperRequest;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Column;
+import jakarta.persistence.Column;
 
 @Entity
 @Table(name = "publisher")
